// src/types_interfaces/enums/gameObjects.ts
export enum CityStreetObjects {
  barrelClosed = "barrelClosed",
  barrelOpened = "barrelOpened",
  barrelBroken = "barrelBroken",
  barrelDestroyed = "barrelDestroyed",
  bench = "bench",
  boxWood1 = "boxWood1",
  boxWood2 = "boxWood2",
  boxPaperSm = "boxPaperSm",
  boxPaperLg = "boxPaperLg",
  boxPaperOpen = "boxPaperOpen",
  boxPaperDamaged = "boxPaperDamaged",
  boxPaperDestroyed = "boxPaperDestroyed",
  boxIron = "boxIron",
  fireEXtinguisher1 = "fireExtinguisher1",
  fireExtinguisher2 = "fireExtinguisher2",
  fireExtinguisher3 = "fireExtinguisher3"
}